chore: split Button struct into separate file
